LayoutTests/http/tests/navigation/page-cache-domcache-pending-promise-expected.txt

Tests that a page with pending DOMCache activity goes into the page cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page was restored from Page Cache
PASS Cache.add() succeeded
PASS !!restoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
